CHORE: cleaned up console.logs codes, moved login actions to it's own file, added PrivateRoute HOC
